Changed gravity/updated collectibles
Now, the Hippo and the food assets have RigidBody2D and BoxCollider2D
components attached. Also, I removed the script that changes the gravity
and instead set the gravity to zero in the food objects and used
velocity to move them.
Lastly, I added a function in PlayerController script to collect
PickUps. However, if the player collects the original food object
instead of its clones, then none of the clones show up.